+++ src/domains.js
@@ -16,15 +16,15 @@ import {
 import { URI } from "./config";
 
 export function getPayPalLoggerDomain(): string {
-  if (__ENV__ === ENV.LOCAL) {
-    const stageHost = getStageHost();
+  // if (__ENV__ === ENV.LOCAL) {
+  //   const stageHost = getStageHost();
 
-    if (!stageHost) {
-      throw new Error(`No stage host found`);
-    }
+  //   if (!stageHost) {
+  //     throw new Error(`No stage host found`);
+  //   }
 
-    return `${getProtocol()}://${stageHost}`;
-  }
+  //   return `${getProtocol()}://${stageHost}`;
+  // }
 
   return getPayPalDomain();
 }
