+++ test/client/common.test.js
@@ -1,9 +1,14 @@
 /* @flow */
-
+import { beforeEach, describe, test } from "vitest";
 import { noop } from "@krakenjs/belter/src";
 
 import { insertMockSDKScript } from "../../src";
-
+describe.skip("", () => {
+  test("", () => {
+    // TODO: move this to a global adding the mock script?
+    // added describe and test to just address this later
+  });
+});
 function clearErrorListener() {
   // eslint-disable-next-line unicorn/prefer-add-event-listener
   window.onerror = noop;
